core-app-api: add missing core feature icons + move icon things around a bit
