+++ packages/core-app-api/src/app/AppManager.test.tsx
@@ -24,7 +24,6 @@ import { lightTheme } from '@backstage/theme';
 import { render, screen } from '@testing-library/react';
 import React, { PropsWithChildren } from 'react';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
-import { defaultAppIcons } from './icons';
 import {
   configApiRef,
   createApiFactory,
@@ -38,7 +37,7 @@ import {
   analyticsApiRef,
 } from '@backstage/core-plugin-api';
 import { generateBoundRoutes, AppManager } from './AppManager';
-import { AppComponents } from './types';
+import { AppComponents, AppIcons } from './types';
 
 describe('generateBoundRoutes', () => {
   it('runs happy path', () => {
@@ -187,6 +186,8 @@ describe('Integration Test', () => {
     ThemeProvider: ({ children }) => <>{children}</>,
   };
 
+  const icons = {} as AppIcons;
+
   it('runs happy paths', async () => {
     const app = new AppManager({
       apis: [noOpAnalyticsApi],
@@ -199,7 +200,7 @@ describe('Integration Test', () => {
           theme: lightTheme,
         },
       ],
-      icons: defaultAppIcons,
+      icons,
       plugins: [],
       components,
       bindRoutes: ({ bind }) => {
@@ -253,7 +254,7 @@ describe('Integration Test', () => {
           theme: lightTheme,
         },
       ],
-      icons: defaultAppIcons,
+      icons,
       plugins: [],
       components,
       bindRoutes: ({ bind }) => {
@@ -310,7 +311,7 @@ describe('Integration Test', () => {
           theme: lightTheme,
         },
       ],
-      icons: defaultAppIcons,
+      icons,
       plugins: [
         createPlugin({
           id: 'test',
@@ -360,7 +361,7 @@ describe('Integration Test', () => {
           theme: lightTheme,
         },
       ],
-      icons: defaultAppIcons,
+      icons,
       plugins: [],
       components,
       bindRoutes: ({ bind }) => {
@@ -420,7 +421,7 @@ describe('Integration Test', () => {
           theme: lightTheme,
         },
       ],
-      icons: defaultAppIcons,
+      icons,
       plugins: [],
       components,
       bindRoutes: ({ bind }) => {
