+++ packages/core-app-api/src/app/index.ts
@@ -17,5 +17,4 @@
 export { createApp } from './createApp';
 export { createSpecializedApp } from './createSpecializedApp';
 export { defaultConfigLoader } from './defaultConfigLoader';
-export type { AppIcons } from './icons';
 export * from './types';
